add logo component
