--- packages/vue-dot/src/patterns/HeaderBar/HeaderBrandSection/index.ts
@@ -1,3 +1,0 @@
-import HeaderBrandSection from './HeaderBrandSection.vue';
-
-export default HeaderBrandSection;
