--- packages/vue-dot/src/patterns/HeaderBar/HeaderBrandSection/types.d.ts
@@ -1,14 +1,0 @@
-export interface LogoInfo {
-	src: string;
-	alt: string;
-}
-
-interface HighlightedTitle {
-	text: string;
-	highlight: string;
-}
-
-export interface Service {
-	title: string | HighlightedTitle;
-	subTitle: string;
-}
