+++ packages/vue-dot/src/patterns/LogoBrandSection/ThemeEnum.ts
@@ -1,0 +1,10 @@
+export enum ThemeEnum {
+	DEFAULT = 'default',
+	CNAM = 'cnam',
+	AMELI = 'ameli',
+	AMELI_PRO = 'ameli-pro',
+	RISQUE_PRO = 'risque-pro',
+	COMPTE_ENTREPRISE = 'compte-entreprise'
+}
+
+export const THEME_ENUM_VALUES = Object.values(ThemeEnum);
