+++ packages/vue-dot/src/patterns/LogoBrandSection/index.ts
@@ -1,0 +1,3 @@
+import LogoBrandSection from './LogoBrandSection.vue';
+
+export default LogoBrandSection;
