+++ packages/vue-dot/src/patterns/LogoBrandSection/locales.ts
