+++ packages/vue-dot/src/patterns/LogoBrandSection/secondaryLogoMapping.ts
@@ -1,21 +1,21 @@
-import { IndexedObject } from '../../../types';
+import { IndexedObject } from '../../types';
 
-import { ThemeEnum } from '../ThemeEnum';
+import { ThemeEnum } from './ThemeEnum';
 
 import { locales } from './locales';
 import { LogoInfo } from './types';
 
 export const secondaryLogoMapping: IndexedObject<LogoInfo> = {
 	[ThemeEnum.CNAM]: {
-		src: require('../../../assets/images/cnam.svg'),
+		src: require('../../assets/images/cnam.svg'),
 		alt: locales.logoCnam
 	},
 	[ThemeEnum.AMELI]: {
-		src: require('../../../assets/images/ameli.svg'),
+		src: require('../../assets/images/ameli.svg'),
 		alt: locales.logoAmeli
 	},
 	[ThemeEnum.AMELI_PRO]: {
-		src: require('../../../assets/images/ameli-pro.svg'),
+		src: require('../../assets/images/ameli-pro.svg'),
 		alt: locales.logoAmeliPro
 	}
 };
