+++ packages/vue-dot/src/patterns/LogoBrandSection/tests/LogoBrandSection.spec.ts
@@ -4,10 +4,10 @@ import { Wrapper } from '@vue/test-utils';
 import { mountComponent } from '@/tests';
 import { html } from '@/tests/utils/html';
 
-import HeaderBrandSection from '../';
-import { ThemeEnum } from '../../ThemeEnum';
+import HeaderBrandSection from '..';
+import { ThemeEnum } from '../ThemeEnum';
 
-import Logo from '../../../../elements/Logo';
+import Logo from '../../../elements/Logo';
 
 Vue.component('Logo', Logo);
